chore: Move RelationshipUtils in test package [DHIS2-17713]
