+++ dhis-2/dhis-support/dhis-support-test/pom.xml
@@ -31,10 +31,6 @@
       <groupId>org.hisp.dhis</groupId>
       <artifactId>dhis-support-external</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.hisp.dhis</groupId>
-      <artifactId>dhis-support-commons</artifactId>
-    </dependency>
 
     <!-- Other -->
     <dependency>
